chore: balance load (#1001)
* chore: balance load

* fix: load

* chore: split long templates test
